rkusbphy: Fixed interrupt registration.
Previously the code was attempting to register interrupts from the
parent node which is useless because it doesn't have any
interrupt/interrupt-names attributes. Now it runs the registration
methods on the OTG and host port child nodes which do have the
attributes.

However, I have never seen the interrupts fire even after this fix so
there is probably something else I am missing.

Also, started to make use of the PHY subsystem but it seems like much of
the initialization will need to move to the _enable method.